gitproj/backend/auth_service/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg[binary]==3.2.13
drf-yasg==1.21.7
python-dotenv==1.0.0
